Extract OrcIcebergIds utility class
